+++ components/tredly-libs/python-common/includes/util.py
@@ -17,11 +17,11 @@ from subprocess import Popen, PIPE
 def findTredlyFile(path):
     # strip trailing slash from path
     path = path.rstrip('/')
-    
+
     # make sure dir exists
     if (not os.path.isdir(path)):
         return None
-    
+
     # look for supported tredly files
     if (os.path.isfile(path + "/tredly.yaml")):    # look for tredly.yaml
         return(path + "/tredly.yaml")
@@ -29,26 +29,26 @@ def findTredlyFile(path):
         return(path + "/Tredlyfile")
     elif (os.path.isfile(path + "/tredly.json")):    # look for tredly.json
         return(path + "/tredly.json")
-    
+
     return None
 
 # copies files or folders from a given source to destination
 # handles /path/to/folder and partition/path/to/folder for source
 def copyFromContainerOrPartition(src, dest, partitionName, chmod = None):
-    
+
     if (re.match('^partition/', src)):  # matches partition
         # create the path to the source file/directory
         source = TREDLY_PARTITIONS_MOUNT + "/" + partitionName + "/" + TREDLY_PTN_DATA_DIR_NAME + "/" + src.split('/', 1)[-1].rstrip('/')
     if (re.match('^/', src)):           # matches container
         # create the path to the source file/directory
         source = builtins.tredlyFile.fileLocation + src.rstrip('/')
-    
+
     # make sure the dest exists
     if (not os.path.isdir(dest)):
         os.makedirs(dest)
-    
+
     dest = dest.rstrip('/') + "/"
-    
+
     # Copy the data in
     cmd = ['cp', '-R', source, dest]
     process = Popen(cmd,  stdin=PIPE, stdout=PIPE, stderr=PIPE)
@@ -58,10 +58,10 @@ def copyFromContainerOrPartition(src, dest, partitionName, chmod = None):
         print(stdErr)
         return False
     else:
-        # check if we need to change the mode 
+        # check if we need to change the mode
         if (chmod is not None):
             os.chmod(dest, chmod)
-        
+
         # Success
         return True
 
@@ -73,28 +73,28 @@ def getAvailableIP4Address(networkAddr, cidr):
     # if cidr == 32 then the host is the network
     if (cidr == 32):
         return networkAddr
-    
+
     # get an ipnetwork object
     network = ip_network(networkAddr + "/" + cidr)
-    
+
     # get a list of hosts and select one at random
     hosts = list(network.hosts())
-    
+
     foundIP = ''
     # loop until we find one thats not in use
     while (len(foundIP) == 0):
         endRand = len(list(hosts)) -1
-        
+
         index = random.randint(0, endRand)
-        
+
         # assign it
         ipAddress = str(hosts[index])
-        
+
         # now make sure its actually unique
         if (not ip4InUse(ipAddress)):
             foundIP = ipAddress
 
-        
+
     return foundIP
 
 def ip4InUse(ip4):
@@ -105,34 +105,34 @@ def ip4InUse(ip4):
     for usedIp4 in inUse:
         if (ip4 == str(usedIp4)):
             return True
-    
+
     return False
 
-    
+
 
 # get a list of ip addresses that are actually in use
 def getIP4AddressesInUse():
     # get a list of ip4_addrs
     zfsPartitions = ZFSDataset(ZFS_TREDLY_DATASET)
-    
+
     # get a list of ip addresses for this interface
     ip4Addrs = zfsPartitions.getPropertyRecursive(ZFS_PROP_ROOT + ':ip4_addr')
-    
+
     # a list to return
     ip4s = []
-    
+
     # loop over results
     for ip4Addr in ip4Addrs:
         #match bridge0|192.168.0.1/24
         regex = '^(\w+)\|([\w.]+)\/(\d+)$'
-        
+
         m = re.match(regex, ip4Addr)
         ip4 = m.group(2)
         # make sure its a valid ip
         if (isValidIp4(ip4)):
             # create an ipv4 address and return
             ip4s.append(IPv4Address(ip4))
-    
+
     return ip4s
 
 
@@ -147,26 +147,26 @@ def getInterfaceIP4(interface):
 def nginxFormatFilename(filename):
     # swap :// for dash
     filename = filename.replace('://', '-')
-    
+
     # swap dots for underscores
     filename = filename.replace('.', '_')
-    
+
     # swap slashes for dashes
     filename = filename.replace('/', '-')
-    
+
     return filename
 
 # formats a filename for unbound
 def unboundFormatFilename(filename):
     # base unbound filename off last 3 (sub)domains
     fileParts = filename.split('.')
-    
+
     # create the filename
     if (len(fileParts) >= 3):
         filename = fileParts[-3] + '.' + fileParts[-2] + '.' + fileParts[-1]
     elif(len(fileParts) == 2):
         filename = fileParts[-2] + '.' + fileParts[-1]
-    
+
     return filename
 
 # checks whether an interface exists on the host or not
@@ -174,9 +174,9 @@ def networkInterfaceExists(interface):
     # if the interface is none then its invalid
     if (interface is None):
         return False
-    
+
     ifconfig = ['ifconfig', interface]
-    
+
     process = Popen(ifconfig, stdin=PIPE, stdout=PIPE, stderr=PIPE)
     stdOut, stdErr = process.communicate()
     if (process.returncode != 0):
@@ -184,13 +184,28 @@ def networkInterfaceExists(interface):
     else:
         return True
 
+# checks whether a hostname is valid or not
+def isValidHostname(hostname):
+    # make sure length isnt > 255 chars
+    if (len(hostname) > 255):
+        return False
+
+    # valid hostname pattern
+    pattern = '^([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]{0,61}[a-zA-Z0-9])(\.([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]{0,61}[a-zA-Z0-9]))*$'
+
+    # check if it matches
+    if (re.match(pattern, hostname)):
+        return True
+
+    return False
+
 # checks whether an ip address is valid or not
 def isValidIp4(ip4):
     try:
         IPv4Address(ip4)
     except:
         return False
-    
+
     return True
 
 # check if a cidr is valid or not
@@ -200,10 +215,10 @@ def isValidCidr(cidr):
         cidrInt = int(cidr)
     except:
         return False
-    
+
     if (cidrInt < 0) or (cidrInt > 32):
         return False
-    
+
     return True
 
 # convert an ip4 address to ip6
@@ -221,7 +236,7 @@ def ip4ToIP6(ip4):
 def validateIp4Addr(ip4Addr):
     regex = '^(\w+)\|([\w.]+)\/(\d+)$'
     m = re.match(regex, ip4Addr)
-    
+
     if (m is None):
         e_error('Could not validate ip4_addr "' + ip4Addr + '" - incorrect format')
         return False
@@ -230,7 +245,7 @@ def validateIp4Addr(ip4Addr):
         interface = m.group(1)
         ip4 = m.group(2)
         cidr = m.group(3)
-        
+
         # make sure the interface exists
         if (not networkInterfaceExists(interface)):
             e_error('Interface "' + interface + '" does not exist in "' + ip4Addr +'"')
@@ -248,5 +263,5 @@ def validateIp4Addr(ip4Addr):
         if (ip4InUse(ip4)):
             e_error("IP Address " + ip4 + ' is already in use.')
             return False
-        
+
     return True